Pull SingleValue functionality out and let all nodes inherit from it
This pulled in an edge case where function nodes use `self` to store data, which requires a new field to store it in to bypass the getattr node injection. I actually like this, since adding state to function nodes in this way was something to discourage to start with!
